Add DisplayContext to the EvaluationListener
This allows the current display context to be used to format
FSharpTypes correctly with respect to the current open module/namespace
set.

Temporarily moved the EvaluationListener notify to the
EvalParsedDefinitions function as there’s an issue where the current
open set of namespace/modules is not in the current state at the point
the generatedValueLayout continuation is constructed.